+++ migrations/2026-06-07-233349-0000_modified-attribute-total-function/down.sql
@@ -1,0 +1,1 @@
+drop function data.modified_attribute_total;