+++ mmoldb-ingest/src/modifier_effects_value.rs
@@ -36,7 +36,7 @@ pub fn update_modifier_effects_values(
                                 valid_from: effects_config.valid_from.naive_utc(),
                                 valid_until: effects_config.valid_until.as_ref().map(DateTime::naive_utc),
                                 attribute: taxa.attribute_id(*effect_attribute),
-                                bonus_type: taxa.effect_type_id(effects_config.bonus_type),
+                                effect_type: taxa.effect_type_id(effects_config.bonus_type),
                                 value: *effect_value,
                             }
                         })